Instanciate ImageManager only when needed
